+++ workbench/app/Candidate.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Workbench\App;
+
+use Maartenpaauw\Specifications\HasSpecifications;
+
+final class Candidate
+{
+    use HasSpecifications;
+}